Adding support to gemini-2.0-flash model (#276)
* Adding support to gemini-2.0-flash model

* Re-organize test structure

* Add GeminiConversation tests

* Add back skip mark to API tests

* line length

* google style key

* fix docstring

* basic dev test for checking functionality

---------
