+++ pyproject.toml
@@ -63,6 +63,7 @@ colorcet = "^3.1.0"
 langchain-anthropic = "^0.1.22"
 anthropic = "^0.33.0"
 docstring-parser = "^0.16"
+langchain-google-genai = "~1.0"
 
 [tool.poetry.extras]
 streamlit = ["streamlit"]
